requirements.txt
pandas
numpy
scikit-learn
networkx
matplotlib
librosa
soundfile
yt-dlp==2026.3.17
yt-dlp-ejs==0.8.0
ytmusicapi
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
